[daemonapi] Relocate token creation validation helpers
- Moved `canCreateAccessToken` and `canCreateRefreshToken` from `lib_post_auth.go`
  to `lib_auth.go` for better organization.
