+++ daemon/daemonapi/lib_auth.go
@@ -25,6 +25,36 @@ var (
 	grantJoin = rbac.GrantJoin.String()
 )
 
+// canCreateAccessToken determines whether an access token can be created based
+// on the token type (refresh token) or authentication strategy (UX or User).
+func (a *DaemonAPI) canCreateAccessToken(ctx echo.Context) bool {
+	if s, ok := ctx.Get("token_use").(string); ok && s == "refresh" {
+		return true
+	}
+	strategy := strategyFromContext(ctx)
+	switch strategy {
+	case daemonauth.StrategyUX:
+	case daemonauth.StrategyUser:
+	default:
+		return false
+	}
+	return true
+}
+
+// canCreateRefreshToken determines if a refresh token can be created based
+// on the authentication strategy from the context: It needs StrategyUX or
+// StrategyUser.
+func (a *DaemonAPI) canCreateRefreshToken(ctx echo.Context) bool {
+	strategy := strategyFromContext(ctx)
+	switch strategy {
+	case daemonauth.StrategyUX:
+	case daemonauth.StrategyUser:
+	default:
+		return false
+	}
+	return true
+}
+
 // accessTokenDuration parses a duration string, returning a clamped time.Duration or a default duration if input is nil or empty.
 func (a *DaemonAPI) accessTokenDuration(s *string) (time.Duration, error) {
 	return converters.TDuration{}.TryConvert(s, time.Minute*10, time.Second, time.Hour)
